+++ client/package.json
@@ -20,12 +20,14 @@
     "react-froala-wysiwyg": "^4.1.3",
     "react-icons": "^4.12.0",
     "react-loading-skeleton": "^3.3.1",
+    "react-markdown": "^9.0.1",
     "react-pro-sidebar": "^1.1.0-alpha.1",
     "react-quill": "^2.0.0",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.20.0",
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
+    "react-syntax-highlighter": "^15.5.0",
     "redux-thunk": "^2.4.2",
     "slate": "^0.101.1",
     "slate-history": "^0.100.0",
@@ -59,6 +61,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-syntax-highlighter": "^15.5.11",
     "tailwindcss": "^3.3.5"
   }
 }
