+++ client/src/utils/Slate/components/Leaf.tsx
@@ -1,0 +1,31 @@
+import { chakra, useColorMode } from "@chakra-ui/react";
+import { RenderLeafProps } from "slate-react";
+
+export const Leaf = ({ attributes, children, leaf }: RenderLeafProps) => {
+  const { colorMode, toggleColorMode } = useColorMode();
+
+  if (leaf.bold) {
+    children = <strong>{children}</strong>;
+  }
+
+  //   if (leaf.code) {
+  //     children = (
+  //       <chakra.code
+  //         padding={"3px"}
+  //         backgroundColor={colorMode === "dark" ? "gray.700" : "gray.200"}
+  //         fontSize={"90%"}>
+  //         {children}
+  //       </chakra.code>
+  //     );
+  //   }
+
+  if (leaf.italic) {
+    children = <em>{children}</em>;
+  }
+
+  if (leaf.underline) {
+    children = <u>{children}</u>;
+  }
+
+  return <span {...attributes}>{children}</span>;
+};
